requirements.txt
networkx>=2.8.8
numpy>=1.23.4
matplotlib>=3.6.2
PySimpleGUI>=4.60.4
Pillow>=9.3.0
pyinstaller>=5.13.2
requests>=2.28.2
pygame>=2.2.0
spotipy>=2.22.1
setuptools>=60.2.0
pydot>=1.4.2